Mh/pek 1026 skjemasplitting (#1837)
* refaktor RedigerAvansertBeregningSkjema til AvansertSkjema og sub-skjema(er)

* finjusterer tester i AvansertSkjemaForAndreBrukere og fjernet restanser av ubetinget begrep i beskrivelser

* legger til testid på de to skjemaene og lager test for RedigerAvansertBeregning

* fjerner ubrukte imports
