+++ src/components/AvansertSkjema/AvansertSkjemaForBrukereMedGradertUfoeretrygd/AvansertSkjemaForBrukereMedGradertUfoeretrygd.module.scss.d.ts
@@ -1,0 +1,18 @@
+declare const classNames: {
+  readonly container: "container";
+  readonly container__hasMobilePadding: "container__hasMobilePadding";
+  readonly form: "form";
+  readonly nowrap: "nowrap";
+  readonly textfield: "textfield";
+  readonly select: "select";
+  readonly "navds-select__container": "navds-select__container";
+  readonly spacer: "spacer";
+  readonly spacer__small: "spacer__small";
+  readonly label: "label";
+  readonly label__margin: "label__margin";
+  readonly description: "description";
+  readonly descriptionText: "descriptionText";
+  readonly ingress: "ingress";
+  readonly alertWrapper: "alertWrapper";
+};
+export = classNames;
