+++ src/components/AvansertSkjema/AvansertSkjemaForBrukereMedGradertUfoeretrygd/index.ts
@@ -1,0 +1,1 @@
+export { AvansertSkjemaForBrukereMedGradertUfoeretrygd } from './AvansertSkjemaForBrukereMedGradertUfoeretrygd'
