+++ src/components/AvansertSkjema/FormButtonRow/FormButtonRow.module.scss
@@ -1,4 +1,4 @@
-@use '../../scss/variables';
+@use '../../../scss/variables';
 
 .wrapper {
   margin-bottom: var(--a-spacing-2);
