+++ src/components/AvansertSkjema/FormButtonRow/FormButtonRow.tsx
@@ -3,6 +3,7 @@ import { useIntl } from 'react-intl'
 
 import { Button } from '@navikt/ds-react'
 
+import { AVANSERT_FORM_NAMES } from '../utils'
 import { BeregningContext } from '@/pages/Beregning/context'
 import { useAppSelector } from '@/state/hooks'
 import {
@@ -10,8 +11,6 @@ import {
   selectIsEndring,
 } from '@/state/userInput/selectors'
 import { wrapLogger } from '@/utils/logging'
-
-import { AVANSERT_FORM_NAMES } from './utils'
 
 import styles from './FormButtonRow.module.scss'
