+++ src/components/AvansertSkjema/FormButtonRow/__tests__/FormButtonRow.test.tsx
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest'
 
+import { AVANSERT_FORM_NAMES } from '../../utils'
 import { FormButtonRow } from '../FormButtonRow'
-import { AVANSERT_FORM_NAMES } from '../utils'
 import { fulfilledGetLoependeVedtakLoependeAlderspensjon } from '@/mocks/mockedRTKQueryApiCalls'
 import {
   BeregningContext,
