+++ src/pages/Beregning/BeregningAvansert.tsx
@@ -32,13 +32,11 @@ import {
   selectAarligInntektFoerUttakBeloep,
   selectIsEndring,
   selectLoependeVedtak,
-  selectNedreAldersgrense,
   selectEpsHarPensjon,
   selectEpsHarInntektOver2G,
   selectSivilstand,
   selectUtenlandsperioder,
 } from '@/state/userInput/selectors'
-import { getBrukerensAlderISluttenAvMaaneden } from '@/utils/alder'
 import { logger } from '@/utils/logging'
 
 import styles from './BeregningAvansert.module.scss'
@@ -57,7 +55,6 @@ export const BeregningAvansert: React.FC = () => {
   const aarligInntektFoerUttakBeloep = useAppSelector(
     selectAarligInntektFoerUttakBeloep
   )
-  const nedreAldersgrense = useAppSelector(selectNedreAldersgrense)
 
   const epsHarPensjon = useAppSelector(selectEpsHarPensjon)
   const epsHarInntektOver2G = useAppSelector(selectEpsHarInntektOver2G)
@@ -178,15 +175,7 @@ export const BeregningAvansert: React.FC = () => {
     <>
       {avansertSkjemaModus === 'redigering' && (
         <RedigerAvansertBeregning
-          gaaTilResultat={() => {
-            setAvansertSkjemaModus('resultat')
-            window.scrollTo(0, 0)
-          }}
           vilkaarsproeving={alderspensjon?.vilkaarsproeving}
-          brukerensAlderPlus1Maaned={getBrukerensAlderISluttenAvMaaneden(
-            person,
-            nedreAldersgrense
-          )}
         />
       )}
